requirements.txt
beautifulsoup4==4.9.3
python-dotenv==0.15.0